feat: Add dtos, add models, add entity logic, start implementation models relations.

+++ src/main/java/org/example/dto/CinemaHallDto.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class CinemaHallDto {
     private String name;
-    private short seatsCount;
+    private int seatsCount;
 }
